+++ CHANGELOG.md
@@ -4,7 +4,13 @@ All notable changes to the array-utils project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-<!--## [Unreleased]-->
+## [Unreleased]
+### BREAKING CHANGES
+- sortArrayByObjectProperty no longer accepts 'undefined' as sortable attribute.
+
+### Added
+- 'Date' as a sortable attribute for sortArrayByObjectProperty.
+
 ## [3.2.0] - 2020-05-18
 ### Changed
 - objectsArrayToObject now takes a selector function, using a string as key is deprecated.